feat : input 컴포넌트 focus, countindicator 에러 해결
